added article delete/add

+++ src/articles/articles.controller.ts
@@ -3,14 +3,18 @@ import {
   Controller,
   Delete,
   Get,
+  HttpException,
+  HttpStatus,
   NotFoundException,
   Param,
   Post,
   UploadedFiles,
+  UseGuards,
   UseInterceptors,
 } from '@nestjs/common';
 import { FilesInterceptor } from '@nestjs/platform-express';
 import { diskStorage } from 'multer';
+import { JwtAuthGuard } from 'src/auth/guards/jwt-auth.guard';
 import { getFilePath, PUBLIC_PATHS } from 'src/constants/public-data-paths';
 import {
   imageMinification,
@@ -18,9 +22,10 @@ import {
   imagesFileName,
 } from 'src/utils/files-helpers';
 import { ArticleService } from './article.service';
-import { ArticleRequestItemInterface } from './dto/article-request-item.interface';
-import { NewArticle } from './dto/new-article.interface';
 import { Article } from './models/article.model';
+import { CurrentUser } from 'src/users/current-user.decorator';
+import { User } from 'src/users/models/user.model';
+import { ArticleRequestItemInterface, NewArticleInput } from './article-dtos';
 
 @Controller()
 export class ArticlesController {
@@ -29,9 +34,8 @@ export class ArticlesController {
   @Get('/articles')
   async getArticles(): Promise<ArticleRequestItemInterface[]> {
     const articles = await this._articleService.findAll();
-
     if (articles) {
-      return articles.map((article) => this.getFullArticle(article));
+      return articles.map((article) => this._getFullArticle(article));
     } else {
       new NotFoundException();
     }
@@ -43,7 +47,7 @@ export class ArticlesController {
   ): Promise<ArticleRequestItemInterface> {
     const article = await this._articleService.findById(parseInt(id));
     if (article) {
-      const fullArticle = this.getFullArticle(article);
+      const fullArticle = this._getFullArticle(article);
       return fullArticle;
     } else {
       new NotFoundException();
@@ -51,39 +55,51 @@ export class ArticlesController {
   }
 
   @Post('/articles/create')
+  @UseGuards(JwtAuthGuard)
   @UseInterceptors(
     FilesInterceptor('image', 1, {
       fileFilter: imagesFileFilter,
       storage: diskStorage({
-        destination: PUBLIC_PATHS.articles,
+        destination: 'public' + PUBLIC_PATHS.articles,
         filename: imagesFileName,
       }),
     }),
   )
   public async createArticle(
-    @Body() newArticleData: NewArticle,
+    @Body() newArticleData: NewArticleInput,
     @UploadedFiles() files: Express.Multer.File[],
-  ) {
+    @CurrentUser() user: User,
+  ): Promise<ArticleRequestItemInterface> {
     await imageMinification(
-      [PUBLIC_PATHS.articles + files[0].filename],
-      PUBLIC_PATHS.articlesThumbnail,
+      ['public' + PUBLIC_PATHS.articles + files[0].filename],
+      'public' + PUBLIC_PATHS.articlesThumbnail,
       [0.6, 0.8],
     );
 
-    await this._articleService.add({
+    const article = await this._articleService.add({
       ...newArticleData,
       image: files[0].filename,
+      userId: user.id,
     });
 
-    return 'ok';
+    return this._getFullArticle(article);
   }
 
   @Delete('/articles/:id')
-  public async removeArticle(@Param('id') id: string) {
+  @UseGuards(JwtAuthGuard)
+  public async removeArticle(
+    @Param('id') id: string,
+    @CurrentUser() user: User,
+  ) {
+    const article = await this._articleService.findById(parseInt(id));
+    if (article.userId !== user.id) {
+      throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
+    }
+
     return await this._articleService.remove(parseInt(id));
   }
 
-  private getFullArticle(article: Article): ArticleRequestItemInterface {
+  private _getFullArticle(article: Article): ArticleRequestItemInterface {
     return {
       ...article,
       imageUrl: getFilePath(PUBLIC_PATHS.articles, article.image),
